Moves dark mode toggle to wrapper.
